+++ guillotina_elasticsearch/exceptions.py
@@ -1,0 +1,5 @@
+from aiohttp.web_exceptions import HTTPException
+
+
+class QueryErrorException(HTTPException):
+    status_code = 488